Let's try not creating the log dir until needed
